requirements.txt
google-generativeai
flask
python-dotenv
PyMuPDF